+++ src/main/java/com/back/moyeomoyeo/repository/member/MemberRepositoryCustom.java
@@ -1,63 +1,13 @@
 package com.back.moyeomoyeo.repository.member;
 
 import com.back.moyeomoyeo.entity.member.Member;
-import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
 
-import static com.back.moyeomoyeo.entity.member.QMember.member;
+public interface MemberRepositoryCustom {
+    boolean existsMemberLoginIdOrNickname(String loginId, String nickname);
 
-@Repository
-@RequiredArgsConstructor
-public class MemberRepositoryCustom {
-    private final JPAQueryFactory queryFactory;
+    boolean existsLoginId(String loginId);
 
-    public boolean existsMemberLoginIdOrNickname(String loginId, String nickname) {
-        Member findMember = queryFactory
-                .selectFrom(member)
-                .where(loginIdEq(loginId).or(nicknameEq(nickname)))
-                .fetchFirst();
-        return findMember != null;
-    }
+    boolean existsNickname(String nickname);
 
-    public boolean existsLoginId(String loginId) {
-
-        Member findMember = queryFactory
-                .selectFrom(member)
-                .where(member.loginId.eq(loginId))
-                .fetchFirst();
-
-        return findMember != null;
-    }
-
-    public boolean existsNickname(String nickname) {
-
-        Member findMember = queryFactory
-                .selectFrom(member)
-                .where(member.nickname.eq(nickname))
-                .fetchFirst();
-
-        return findMember != null;
-    }
-
-    public Member findByNickname(String nickname) {
-        return queryFactory
-                .selectFrom(member)
-                .where(nicknameEq(nickname))
-                .fetchOne();
-    }
-
-
-    private BooleanExpression loginIdEq(String loginId) {
-        return loginId != null ? member.loginId.eq(loginId) : null;
-    }
-
-    private BooleanExpression usernameEq(String username) {
-        return username != null ? member.username.eq(username) : null;
-    }
-
-    private BooleanExpression nicknameEq(String nickname) {
-        return nickname != null ? member.nickname.eq(nickname) : null;
-    }
+    Member findByNickname(String nickname);
 }
